# backend/openapi.json
{"openapi":"3.0.3","info":{"title":"Smart Room API","version":"1.0.0"},"components":{"schemas":{"LoginBody":{"type":"object","properties":{"email":{"type":"string","format":"email","pattern":"^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"},"password":{"type":"string","minLength":1}},"required":["email","password"]},"CreateBuildingBody":{"type":"object","properties":{"name":{"type":"string","minLength":1,"maxLength":100}},"required":["name"]},"UpdateBuildingBody":{"type":"object","properties":{"name":{"type":"string","minLength":1,"maxLength":100}},"required":["name"]},"CreateFloorBody":{"type":"object","properties":{"name":{"type":"string","minLength":1,"maxLength":100},"building_id":{"type":"integer","minimum":0,"exclusiveMinimum":true,"maximum":9007199254740991}},"required":["name","building_id"]},"UpdateFloorBody":{"type":"object","properties":{"name":{"type":"string","minLength":1,"maxLength":100},"building_id":{"type":"integer","minimum":0,"exclusiveMinimum":true,"maximum":9007199254740991}},"required":["name","building_id"]},"CreateReservationBody":{"type":"object","properties":{"room_id":{"type":"integer","minimum":0,"exclusiveMinimum":true,"maximum":9007199254740991},"start_time":{"type":"string"},"end_time":{"type":"string"}},"required":["room_id","start_time","end_time"]},"UpdateReservationBody":{"type":"object","properties":{"room_id":{"type":"integer","minimum":0,"exclusiveMinimum":true,"maximum":9007199254740991},"start_time":{"type":"string"},"end_time":{"type":"string"}},"required":["room_id","start_time","end_time"]},"CreateRoomBody":{"type":"object","properties":{"name":{"type":"string","minLength":1,"maxLength":100},"floor_id":{"type":"integer","minimum":0,"exclusiveMinimum":true,"maximum":9007199254740991},"capacity":{"type":"integer","minimum":0,"exclusiveMinimum":true,"maximum":9007199254740991}},"required":["name","floor_id","capacity"]},"UpdateRoomBody":{"type":"object","properties":{"name":{"type":"string","minLength":1,"maxLength":100},"floor_id":{"type":"integer","minimum":0,"exclusiveMinimum":true,"maximum":9007199254740991},"capacity":{"type":"integer","minimum":0,"exclusiveMinimum":true,"maximum":9007199254740991}},"required":["name","floor_id","capacity"]},"CreateUserBody":{"type":"object","properties":{"email":{"type":"string","format":"email","pattern":"^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"},"password":{"type":"string","minLength":8},"first_name":{"type":"string","minLength":1,"maxLength":100},"last_name":{"type":"string","minLength":1,"maxLength":100}},"required":["email","password","first_name","last_name"]},"User":{"type":"object","properties":{"id":{"type":"integer","minimum":-9007199254740991,"maximum":9007199254740991},"last_name":{"type":"string"},"first_name":{"type":"string"},"role":{"$ref":"#/components/schemas/Role"},"email":{"type":"string","format":"email","pattern":"^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"}},"required":["id","last_name","first_name","role","email"],"additionalProperties":false},"Role":{"type":"string","enum":["USER","ADMIN"]},"InvalidBodyError":{"type":"object","properties":{"code":{"type":"string","enum":["INVALID_BODY"]},"issues":{}},"required":["code","issues"],"additionalProperties":false},"InvalidCredentialsError":{"type":"object","properties":{"code":{"type":"string","enum":["INVALID_CREDENTIALS"]},"message":{"type":"string"}},"required":["code","message"],"additionalProperties":false},"InternalServerError":{"type":"object","properties":{"code":{"type":"string","enum":["INTERNAL_SERVER_ERROR"]},"message":{"type":"string"}},"required":["code","message"],"additionalProperties":false},"UnauthorizedError":{"type":"object","properties":{"code":{"type":"string","enum":["UNAUTHORIZED"]},"message":{"type":"string"}},"required":["code","message"],"additionalProperties":false},"Building":{"type":"object","properties":{"id":{"type":"integer","minimum":-9007199254740991,"maximum":9007199254740991},"name":{"type":"string"}},"required":["id","name"],"additionalProperties":false},"BuildingNameExistsError":{"type":"object","properties":{"code":{"type":"string","enum":["BUILDING_NAME_EXISTS"]},"message":{"type":"string"}},"required":["code","message"],"additionalProperties":false},"InvalidParamsError":{"type":"object","properties":{"code":{"type":"string","enum":["INVALID_PARAMS"]},"issues":{}},"required":["code","issues"],"additionalProperties":false},"BuildingNotFoundError":{"type":"object","properties":{"code":{"type":"string","enum":["BUILDING_NOT_FOUND"]},"message":{"type":"string"}},"required":["code","message"],"additionalProperties":false},"BuildingHasFloorsError":{"type":"object","properties":{"code":{"type":"string","enum":["BUILDING_HAS_FLOORS"]},"message":{"type":"string"}},"required":["code","message"],"additionalProperties":false},"Floor":{"type":"object","properties":{"id":{"type":"integer","minimum":-9007199254740991,"maximum":9007199254740991},"name":{"type":"string"},"building_id":{"type":"integer","minimum":-9007199254740991,"maximum":9007199254740991}},"required":["id","name","building_id"],"additionalProperties":false},"FloorNameExistsError":{"type":"object","properties":{"code":{"type":"string","enum":["FLOOR_NAME_EXISTS"]},"message":{"type":"string"}},"required":["code","message"],"additionalProperties":false},"FloorNotFoundError":{"type":"object","properties":{"code":{"type":"string","enum":["FLOOR_NOT_FOUND"]},"message":{"type":"string"}},"required":["code","message"],"additionalProperties":false},"FloorHasRoomsError":{"type":"object","properties":{"code":{"type":"string","enum":["FLOOR_HAS_ROOMS"]},"message":{"type":"string"}},"required":["code","message"],"additionalProperties":false},"Reservation":{"type":"object","properties":{"id":{"type":"integer","minimum":-9007199254740991,"maximum":9007199254740991},"room_id":{"nullable":true,"type":"integer","minimum":-9007199254740991,"maximum":9007199254740991},"user_id":{"nullable":true,"type":"integer","minimum":-9007199254740991,"maximum":9007199254740991},"start_time":{"type":"string"},"end_time":{"type":"string"}},"required":["id","room_id","user_id","start_time","end_time"],"additionalProperties":false},"InvalidDateRangeError":{"type":"object","properties":{"code":{"type":"string","enum":["INVALID_DATE_RANGE"]},"message":{"type":"string"}},"required":["code","message"],"additionalProperties":false},"RoomNotFoundError":{"type":"object","properties":{"code":{"type":"string","enum":["ROOM_NOT_FOUND"]},"message":{"type":"string"}},"required":["code","message"],"additionalProperties":false},"RoomNotAvailableError":{"type":"object","properties":{"code":{"type":"string","enum":["ROOM_NOT_AVAILABLE"]},"message":{"type":"string"}},"required":["code","message"],"additionalProperties":false},"ReservationNotFoundError":{"type":"object","properties":{"code":{"type":"string","enum":["RESERVATION_NOT_FOUND"]},"message":{"type":"string"}},"required":["code","message"],"additionalProperties":false},"Room":{"type":"object","properties":{"id":{"type":"integer","minimum":-9007199254740991,"maximum":9007199254740991},"name":{"type":"string"},"floor_id":{"type":"integer","minimum":-9007199254740991,"maximum":9007199254740991},"capacity":{"type":"integer","minimum":-9007199254740991,"maximum":9007199254740991}},"required":["id","name","floor_id","capacity"],"additionalProperties":false},"RoomNameExistsError":{"type":"object","properties":{"code":{"type":"string","enum":["ROOM_NAME_EXISTS"]},"message":{"type":"string"}},"required":["code","message"],"additionalProperties":false},"InvalidQueryError":{"type":"object","properties":{"code":{"type":"string","enum":["INVALID_QUERY"]},"issues":{}},"required":["code","issues"],"additionalProperties":false},"EmailAlreadyExistsError":{"type":"object","properties":{"code":{"type":"string","enum":["EMAIL_ALREADY_EXISTS"]},"message":{"type":"string"}},"required":["code","message"],"additionalProperties":false},"UserNotFoundError":{"type":"object","properties":{"code":{"type":"string","enum":["USER_NOT_FOUND"]},"message":{"type":"string"}},"required":["code","message"],"additionalProperties":false}}},"paths":{"/auth/login":{"post":{"tags":["Auth"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginBody"}}}},"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"400":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvalidBodyError"}}}},"401":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvalidCredentialsError"}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}}}}}},"/auth/logout":{"post":{"tags":["Auth"],"responses":{"204":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{},"additionalProperties":false}}}},"401":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedError"}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}}}}}},"/auth/logout-all":{"post":{"tags":["Auth"],"responses":{"204":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{},"additionalProperties":false}}}},"401":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedError"}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}}}}}},"/buildings":{"post":{"tags":["Buildings"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBuildingBody"}}}},"responses":{"201":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Building"}}}},"400":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvalidBodyError"}}}},"401":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedError"}}}},"409":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuildingNameExistsError"}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}}}}},"get":{"tags":["Buildings"],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Building"}}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}}}}}},"/buildings/{id}":{"get":{"tags":["Buildings"],"parameters":[{"schema":{"type":"integer","minimum":0,"exclusiveMinimum":true,"maximum":9007199254740991},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Building"}}}},"400":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvalidParamsError"}}}},"404":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuildingNotFoundError"}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}}}}},"put":{"tags":["Buildings"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateBuildingBody"}}}},"parameters":[{"schema":{"type":"integer","minimum":0,"exclusiveMinimum":true,"maximum":9007199254740991},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Building"}}}},"400":{"description":"Default Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/InvalidParamsError"},{"$ref":"#/components/schemas/InvalidBodyError"}]}}}},"401":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedError"}}}},"404":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuildingNotFoundError"}}}},"409":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuildingNameExistsError"}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}}}}},"delete":{"tags":["Buildings"],"parameters":[{"schema":{"type":"integer","minimum":0,"exclusiveMinimum":true,"maximum":9007199254740991},"in":"path","name":"id","required":true}],"responses":{"204":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{},"additionalProperties":false}}}},"400":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvalidParamsError"}}}},"401":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedError"}}}},"404":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuildingNotFoundError"}}}},"409":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuildingHasFloorsError"}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}}}}}},"/floors":{"post":{"tags":["Floors"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateFloorBody"}}}},"responses":{"201":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Floor"}}}},"400":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvalidBodyError"}}}},"401":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedError"}}}},"404":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuildingNotFoundError"}}}},"409":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FloorNameExistsError"}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}}}}},"get":{"tags":["Floors"],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Floor"}}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}}}}}},"/floors/{id}":{"get":{"tags":["Floors"],"parameters":[{"schema":{"type":"integer","minimum":0,"exclusiveMinimum":true,"maximum":9007199254740991},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Floor"}}}},"400":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvalidParamsError"}}}},"404":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FloorNotFoundError"}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}}}}},"put":{"tags":["Floors"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateFloorBody"}}}},"parameters":[{"schema":{"type":"integer","minimum":0,"exclusiveMinimum":true,"maximum":9007199254740991},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Floor"}}}},"400":{"description":"Default Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/InvalidParamsError"},{"$ref":"#/components/schemas/InvalidBodyError"}]}}}},"401":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedError"}}}},"404":{"description":"Default Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/FloorNotFoundError"},{"$ref":"#/components/schemas/BuildingNotFoundError"}]}}}},"409":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FloorNameExistsError"}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}}}}},"delete":{"tags":["Floors"],"parameters":[{"schema":{"type":"integer","minimum":0,"exclusiveMinimum":true,"maximum":9007199254740991},"in":"path","name":"id","required":true}],"responses":{"204":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{},"additionalProperties":false}}}},"400":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvalidParamsError"}}}},"401":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedError"}}}},"404":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FloorNotFoundError"}}}},"409":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FloorHasRoomsError"}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}}}}}},"/health":{"get":{"tags":["Health"],"responses":{"204":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{},"additionalProperties":false}}}}}}},"/reservations":{"post":{"tags":["Reservations"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReservationBody"}}}},"responses":{"201":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Reservation"}}}},"400":{"description":"Default Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/InvalidBodyError"},{"$ref":"#/components/schemas/InvalidDateRangeError"}]}}}},"401":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedError"}}}},"404":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomNotFoundError"}}}},"409":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomNotAvailableError"}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}}}}},"get":{"tags":["Reservations"],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Reservation"}}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}}}}}},"/reservations/{id}":{"get":{"tags":["Reservations"],"parameters":[{"schema":{"type":"integer","minimum":0,"exclusiveMinimum":true,"maximum":9007199254740991},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Reservation"}}}},"400":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvalidParamsError"}}}},"404":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationNotFoundError"}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}}}}},"put":{"tags":["Reservations"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateReservationBody"}}}},"parameters":[{"schema":{"type":"integer","minimum":0,"exclusiveMinimum":true,"maximum":9007199254740991},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Reservation"}}}},"400":{"description":"Default Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/InvalidParamsError"},{"$ref":"#/components/schemas/InvalidBodyError"},{"$ref":"#/components/schemas/InvalidDateRangeError"}]}}}},"401":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedError"}}}},"404":{"description":"Default Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/ReservationNotFoundError"},{"$ref":"#/components/schemas/RoomNotFoundError"}]}}}},"409":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomNotAvailableError"}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}}}}},"delete":{"tags":["Reservations"],"parameters":[{"schema":{"type":"integer","minimum":0,"exclusiveMinimum":true,"maximum":9007199254740991},"in":"path","name":"id","required":true}],"responses":{"204":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{},"additionalProperties":false}}}},"400":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvalidParamsError"}}}},"401":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedError"}}}},"404":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationNotFoundError"}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}}}}}},"/users/{userId}/reservations":{"get":{"tags":["Reservations"],"parameters":[{"schema":{"type":"integer","minimum":0,"exclusiveMinimum":true,"maximum":9007199254740991},"in":"path","name":"userId","required":true}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Reservation"}}}}},"400":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvalidParamsError"}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}}}}}},"/rooms/{roomId}/reservations":{"get":{"tags":["Reservations"],"parameters":[{"schema":{"type":"integer","minimum":0,"exclusiveMinimum":true,"maximum":9007199254740991},"in":"path","name":"roomId","required":true}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Reservation"}}}}},"400":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvalidParamsError"}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}}}}}},"/rooms":{"post":{"tags":["Rooms"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoomBody"}}}},"responses":{"201":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Room"}}}},"400":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvalidBodyError"}}}},"401":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedError"}}}},"404":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FloorNotFoundError"}}}},"409":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomNameExistsError"}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}}}}},"get":{"tags":["Rooms"],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Room"}}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}}}}}},"/rooms/available":{"get":{"tags":["Rooms"],"parameters":[{"schema":{"type":"string","format":"date-time","pattern":"^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"},"in":"query","name":"start","required":true},{"schema":{"type":"string","format":"date-time","pattern":"^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"},"in":"query","name":"end","required":true}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Room"}}}}},"400":{"description":"Default Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/InvalidQueryError"},{"$ref":"#/components/schemas/InvalidDateRangeError"}]}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}}}}}},"/rooms/{id}":{"get":{"tags":["Rooms"],"parameters":[{"schema":{"type":"integer","minimum":0,"exclusiveMinimum":true,"maximum":9007199254740991},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Room"}}}},"400":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvalidParamsError"}}}},"404":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomNotFoundError"}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}}}}},"put":{"tags":["Rooms"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoomBody"}}}},"parameters":[{"schema":{"type":"integer","minimum":0,"exclusiveMinimum":true,"maximum":9007199254740991},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Room"}}}},"400":{"description":"Default Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/InvalidParamsError"},{"$ref":"#/components/schemas/InvalidBodyError"}]}}}},"401":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedError"}}}},"404":{"description":"Default Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/RoomNotFoundError"},{"$ref":"#/components/schemas/FloorNotFoundError"}]}}}},"409":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomNameExistsError"}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}}}}},"delete":{"tags":["Rooms"],"parameters":[{"schema":{"type":"integer","minimum":0,"exclusiveMinimum":true,"maximum":9007199254740991},"in":"path","name":"id","required":true}],"responses":{"204":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{},"additionalProperties":false}}}},"400":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvalidParamsError"}}}},"401":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedError"}}}},"404":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomNotFoundError"}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}}}}}},"/users":{"post":{"tags":["Users"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserBody"}}}},"responses":{"201":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"400":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvalidBodyError"}}}},"401":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedError"}}}},"409":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailAlreadyExistsError"}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}}}}}},"/users/me":{"get":{"tags":["Users"],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"401":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedError"}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}}}}}},"/users/{id}":{"get":{"tags":["Users"],"parameters":[{"schema":{"type":"integer","minimum":0,"exclusiveMinimum":true,"maximum":9007199254740991},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"400":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvalidParamsError"}}}},"404":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserNotFoundError"}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerError"}}}}}}}}}